+++ gameserver/src/main/java/com/shnok/javaserver/gameserver/model/actor/Playable.java
@@ -302,6 +302,7 @@ public abstract class Playable extends Creature
 	@Override
 	public void updateEffectIcons(boolean partyOnly)
 	{
+		System.out.println("inside playable");
 		_effects.updateEffectIcons(partyOnly);
 	}
