+++ gameserver/src/main/java/com/shnok/javaserver/gameserver/model/actor/container/player/Punishment.java
@@ -9,7 +9,7 @@ import com.shnok.javaserver.gameserver.enums.PunishmentType;
 import com.shnok.javaserver.gameserver.enums.ZoneId;
 import com.shnok.javaserver.gameserver.model.actor.Player;
 import com.shnok.javaserver.gameserver.model.olympiad.OlympiadManager;
-import com.shnok.javaserver.gameserver.network.serverpackets.unused.EtcStatusUpdate;
+import com.shnok.javaserver.gameserver.network.serverpackets.effect.EtcStatusUpdate;
 import com.shnok.javaserver.gameserver.network.serverpackets.unused.NpcHtmlMessage;
 import com.shnok.javaserver.gameserver.network.serverpackets.unused.PlaySound;
