+++ gameserver/src/main/java/com/shnok/javaserver/gameserver/network/serverpackets/effect/PartySpelled.java
@@ -1,4 +1,4 @@
-package com.shnok.javaserver.gameserver.network.serverpackets.unused;
+package com.shnok.javaserver.gameserver.network.serverpackets.effect;
 
 import java.util.ArrayList;
 import java.util.List;
