+++ gameserver/src/main/java/com/shnok/javaserver/gameserver/network/serverpackets/effect/ShortBuffStatusUpdate.java
@@ -1,4 +1,4 @@
-package com.shnok.javaserver.gameserver.network.serverpackets.unused;
+package com.shnok.javaserver.gameserver.network.serverpackets.effect;
 
 import com.shnok.javaserver.gameserver.network.serverpackets.L2GameServerPacket;
