+++ gameserver/src/main/java/com/shnok/javaserver/gameserver/skills/AbstractEffect.java
@@ -17,9 +17,9 @@ import com.shnok.javaserver.gameserver.model.actor.Player;
 import com.shnok.javaserver.gameserver.model.actor.Summon;
 import com.shnok.javaserver.gameserver.model.actor.instance.Servitor;
 import com.shnok.javaserver.gameserver.network.SystemMessageId;
-import com.shnok.javaserver.gameserver.network.serverpackets.unused.AbnormalStatusUpdate;
+import com.shnok.javaserver.gameserver.network.serverpackets.effect.AbnormalStatusUpdate;
 import com.shnok.javaserver.gameserver.network.serverpackets.unused.ExOlympiadSpelledInfo;
-import com.shnok.javaserver.gameserver.network.serverpackets.unused.PartySpelled;
+import com.shnok.javaserver.gameserver.network.serverpackets.effect.PartySpelled;
 import com.shnok.javaserver.gameserver.network.serverpackets.SystemMessage;
 import com.shnok.javaserver.gameserver.skills.basefuncs.Func;
 import com.shnok.javaserver.gameserver.skills.basefuncs.FuncTemplate;
