+++ gameserver/src/main/java/com/shnok/javaserver/gameserver/skills/effects/EffectCharmOfCourage.java
@@ -4,7 +4,7 @@ import com.shnok.javaserver.gameserver.enums.skills.EffectFlag;
 import com.shnok.javaserver.gameserver.enums.skills.EffectType;
 import com.shnok.javaserver.gameserver.model.actor.Creature;
 import com.shnok.javaserver.gameserver.model.actor.Player;
-import com.shnok.javaserver.gameserver.network.serverpackets.unused.EtcStatusUpdate;
+import com.shnok.javaserver.gameserver.network.serverpackets.effect.EtcStatusUpdate;
 import com.shnok.javaserver.gameserver.skills.AbstractEffect;
 import com.shnok.javaserver.gameserver.skills.L2Skill;
